requirements.txt
pyyaml 
requests
SPARQLWrapper
tqdm
regex 
ujson 
pandas
torch>=2.2
tokenizers>=0.15
numpy>=1.26
